Fixed Routing
Turns out PreactJS does not have a base url system. This lead to the system working off of / instead of /resume/ as it should for github pages. After way too many hours of trying to manipulate the sytem at a system level, (LLMs saying somethign existed but it was for React. Once again another reason to neven use LLMs with problems) changed all internal href links to be relative or hardcoded to /resume/